Gtk3L Reduce vertical space in the CRT controls
When passed `TRUE` as the `minimal` argument, the CRT controls widget will
now use two columns for the sliders and their labels. Which means it now
properly displays on my 1600x900 display with the MATE desktop, which uses
two bars: a main menu bar at the top and a 'task bar' at the bottom (very
much like Gnome2).
The behaviour of the widget when used in the settings UI is still the same,
all sliders are presented one under the other, with default sizes for the
slider indicators and the labels.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@34976 379a1393-f5fb-40a0-bcee-ef074d9b53f7