Add linenoise-ng to src/lib to replace readline
Currently only c1541 uses this, and it sucks with completion. But at least it byulds and links. I had to move files linenoise-ng source files around and create a Makefile.am to be able to create a liblinenoiseng.a, also had to fix a few C++flags in that Makefile as a temporary messure, once we fix configure to not alter C[PP|XX]?FLAGS and LDFLAGS, we need to update the Makefile.am.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@36890 379a1393-f5fb-40a0-bcee-ef074d9b53f7